+++ docs/changelog.md
@@ -1,6 +1,7 @@
 unreleased
 
 - fix FML NPE with translate(), cc(), and c() when assigning to polymorphic elements like value[x] or location[x] (#480)
+- load internal dependencies (ig-internal-dependency extension) from ImplementationGuide resources (#481)
 
 2026/02/12 Release 4.0.18